Don't stop text-decoration-propagation in general for UA shadows.
We used to stop propagation of text-decoration into elements rendered
with UA shadow trees in general. Instead we should list the specific
cases where we need to do this. This is currently for media elements
which has a special rendering where the media controls do not use a
proper containing block model.

This fixes text-decoration propagation for elements like, summary,
details, object, and marquee.

TEST=fast/table/backgr_layers-opacity.html
TEST=fast/table/backgr_layers-opacity-collapsed-border.html
TEST=media/controls-styling.html

Bug: 1042760
Change-Id: Ice2681e4ac93b1a96ec598bcb614b6a8f57e0b21